services/fea-worker/requirements.txt
requests
python-dotenv
pydantic

